requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
deprecation==2.1.0
exceptiongroup==1.2.0
frozendict==2.4.0
gotrue==2.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httpx==0.25.2
idna==3.6
lxml==5.1.0
multitasking==0.0.11
numpy==1.26.4
packaging==23.2
pandas==2.2.1
peewee==3.17.1
postgrest==0.16.1
pydantic==2.6.3
pydantic_core==2.16.3
pyrate-limiter==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.2
requests==2.31.0
requests-ratelimiter==0.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
storage3==0.7.0
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
websockets==11.0.3
yfinance==0.2.37